chore(server): rename controllers
- remove `Controller` from name as it's overly verbose
