+++ server/routes/controllers/forgotPassword.js
